add-parser-tests-1
Add a proper suite of parser tests.  These are basically just the tests in the original,
list-based test-parser function converted to DEFTEST forms.

It's not as transparent as I'd like, since we are actually checking lists generated by
structure-to-plist against other lists rather than checking the structures directly.

Another possibility might be to write a struct= function and write our tests like this:

	(deftest foo/1 :notes parser
	  (struct=
	    (parse "foo null;")
	    #s(foo :arg #s(special-value :symbol :null)))
	  t)

I don't like that quite as much just because when it fails it will give an unhelpful message.

	Expected: T
	Actual: NIL

Gee, thanks. :)

This suite of tests already exposes a problem with the current parser implementation;
test parse-statement-block/1 is failing because it tries to interpret a statement block
as an object literal.

I've only run the suite against parsergen so far.

darcs-hash:20051113021142-53808-c91f2480307a344db3f72050dc56937b53bdd15e