add-test-functionality
This patch adds an implementation of asdf:test-op for the :jwacs system.
So, from a standing start in a fresh lisp image, it should be possible to
say:

  (require :asdf)
  (asdf:oos 'asdf:test-op :jwacs)

and asdf should load the jwacs system, load the jwacs-tests system, and run
all of the tests.

Side note:
Under cl-yacc, it appears that there are 8 failing tests.  The 1 test that
fails under parsergen is not one of them!  Some of them involve var-decl-statements,
so it may be the $10 issue; would you mind looking at them?

darcs-hash:20051113024016-53808-60c81a5fce9eb43678dfaa744a9ee744f69d2e83