yacc-rule-reorder
Okay so I'm an idiot.  I originally thought the rules weren't working because
the output order was different, so I tried some tests and it didn't prove
conclusive. However, turns out I must've done something dumb with those tests
because it *IS* conclusive. The ordering was wrong and that was causing a lot
of the parsing errors. However, we still seem to have 1 parse error left with
Yacc but I'm too tired right now to investigate. I'll get to it shortly.

darcs-hash:20051116025341-2e96c-a8198708f0adcaf97fe0c987e907c71c78da3768