add-cps-transform
Also a "don't get too excited" event. :)

This patch adds the beginnings of a CPS transformation.  The transformation
assumes that the input has been scope-analyzed and "explicitized" (ie, no
implicit results).

Neither of these conditions can currently be guaranteed to be true,
since the transformation to guarantee the first condition is currently under
development, and I haven't even started the transformation to guarantee the
second condition.

With hand-prepared input, this transformation is working on a naive factorial
implementation, as well as a meaningless foo/bar/baz example that exercises
some slightly more-advanced functionality.  I have a new example that is failing,
but I will work on that this weekend sometime.

Other misc changes:
1. Added section levels for header comments
2. got rid of the comment for the COLLAPSE-ADJACENT-VARS transformation, since
   we don't actually need it and it would in fact conflict with the upcoming
   explicitization transformation.
3. Added a couple of TODO comments about future testing

darcs-hash:20051119063841-53808-a1444036cefb627fb62acaee47fea69cb2621857