cps-tests-2
This is the second of three tests of the CPS transformation (it's not actually
a factorial function, hence the different patch name).

"Symmetric dangling tail" just means that the statement-tail for both the if and
the else blocks includes statements from after the if statement (and which therefore
gets "consumed").

The next test (which _is_ a factorial test) will be an assymetric dangling tail,
where only the if case has a statement-tail that includes statements after the
if statement (and where, therefore, the statements after the if statement do not
get "consumed").

darcs-hash:20051121054954-53808-6009229e06c2db9490cca3efd49ac6d6b761adb1