fix-lexer
The lexer wasn't lexing the <= and => tokens because they appeared after < and >
in the list of operators; instead, it was returning < followed by = and similarly.
This checkin reorders the operator list to fix that problem, and adds a new unit
test as a regression check.

darcs-hash:20051127012449-53808-72371e605048dc6d5b2258978bd5fba696d57cb2