fixed-object-literal
Okay, so I didn't really fix it.

I fixed the tests.  Same problem as the function expressions, the object literal needs to be part of an expression, otherwise the "{" gets confused with a statement block.

So, added a var-decl on it and blammo.

darcs-hash:20051128023945-2e96c-d96d3b6a82242f8804547d664ca96ebfc26ac946