explicitization-bugs
I've found some bugs in the explicitization transformation.  This patch
adds a test that reproduces one of them.  It also adds the notion of a
PRIMITIVE-VALUE, which I've used to refine an assertion that is made
during the transformation.

darcs-hash:20051203051630-53808-9e7564a3d538a49d3cbafb1ad741333d3a7fd789