fix-explicitization-compression
The "compression" optimization in the explicitization transformation was
improperly compressing away binary operations.  This was happening because the
EXPLICITIZE-RHS function was assuming that any time it had both some
pre-statements and also an assignment-slot that it was safe to compress away
the final pre-statement.  There was previously an assertion with a long list
of additional conditions that had to be fulfilled before this was a valid
optimization; these extra conditions have been moved out of the assetion
and into the main test.

I've also removed all of the code supporting the notion of primitive values,
since the only purpose for that code was to support a weaker, incorrect
version of the same assertion.

darcs-hash:20051204010420-53808-9d51c71aeb0149647e8dbb9c93d7c88b7b8fadce