add-forbid-transformation-elements
Added a convenience macro named FORBID-TRANSFORMATION-ELEMENTS that
generates DEFMETHOD forms that throw an error if a specified transformation
is applied to any of a list of element types.

This is useful for asserting the absence of certain source-element types.
I've added invocations for both the CPS and EXPLICITIZE transformations
forbidding any of the loop source-element types.

darcs-hash:20051204013420-53808-315fddf3a07db46f4951695f97ca732f3f5bcb39