add-transformation-support-for-object-literals
OBJECT-LITERAL values are a little unusual, because they contain an alist
of cons cells rather than a regular list of lists or structures.  Because
of this weirdness, most of the transformations were choking, because the
standard recursion behaviour doesn't work for alists.

This patch adds support for OBJECT-LITERAL recursion to the standard
transformation recursion behaviour, to the SHIFT-FUNCTION-DECLS
transformation (which defines its own recursion behaviour), and to the
COLLECT-IN-SCOPE function.

It also adds some unit tests to exercise the various transformations in
the presence of OBJECT-LITERAL values.

darcs-hash:20051209220218-53808-6eda5ffcbcef2a41c88ca4cd4e145ca01cea4ce8