trampoline-calls-runtime
We are now more sophisticated about determining (within the trampoline
transformation) the status of return statement arguments.
They can be one of:

1. Tail-call to a definitely-trampolined function, in which case it
   should be wrapped in a thunk.
2. Tail-call to a possibly-trampolined function, in which case it
   should be delegated to the `$trampolineResult` runtime function
   for determination at runtime.
3. Not a function call, in which case it should be set as the `result`
   field of a trampoline result object.

This patch adds the `$trampolineResult` runtime function, which either calls its
argument and returns the result in the `result` field of a "trampoline-boxed"
return value, or wraps its function-call argument in a thunk and returns that.
The decision is made based upon the run-time value of the `$isTrampolined`
property of the function.

darcs-hash:20060102222157-53808-370b0aa7a4dc03d0505e47955326beff61590737