pretty-ugly-print
We need/want to uniquify variable names even when we don't want to obfuscate the
code, so I've added an alternative GENVAR scheme that appends "$n" to the
original identifier name instead of to "JW".

The new scheme will be used whenever *PRETTY-MODE* is T and an original
identifer name is provided; the original scheme will be used in all other
situations.

darcs-hash:20060103050628-53808-90ad42163814c345462c3a1ddd382a45339ac69a