procrastination-refactoring-2
Refactored the environment handling in ugly-print.lisp to make it a little more
idiomatic.  We now use the WITH-ADDED-ENVIRONMENT macro to add a new environment
to the front of the environment stack, and ADD-BINDING now uses PUSH instead of
a SETF/CONS combo.

darcs-hash:20060103051821-53808-a3cf345edf5b8ba781e16642ef4a200f2c0600b7