allow-resume-return-values
The original syntax for the `resume` statement accepted a continuation
but not an argument to pass to that continuation.  That was okay under
the old suspend/resume semantics, because the continuations that `suspend`
produced never took any arguments.

However, the continuations produced by `function_continuation` most
certainly do accept an argument, since that's how we return values
from functions.  Which raises the possibility that we may want to allow
the user to provide a return value when they are resuming their saved
function continuations.

This patch adds syntax to the `resume` statement to allow the user to pass in an
argument to the continuation.  It's pretty ugly, so if you have an idea for a
better syntax I'm all ears.  (The parser seems to get pretty unhappy if there
isn't any separator between the expression specifying the continuation target
and the continuation argument, which is where the `<-` came from).

The following two forms are now supported for `resume`:

	resume foo.bar;		// old style, no argument provided to continuation stored at foo.bar

	resume foo.bar <- 42;  // Pass 42 to continuation stored at foo.bar

darcs-hash:20060108014037-53808-a4e03ffa4b1f7faaddb58233f2a1be694fb103bb