remove-runtime-dependencies
We've had dependencies on the runtime scattered through the CPS and TRAMPOLINE
transformations up until now.  It's made those two transformations kind of
complicated, and it's made it very difficult to change how the runtime is set
up.  So, as a two-stage process, I am moving the dependencies on the runtime
into a single new transformation (the RUNTIME transformation).

This patch is stage 1, where I remove runtime dependencies.  The CPS and
TRAMPOLINE transformations now assume that we are transforming the entire
program and will therefore never have to interact with direct-style functions.
This means, among other things:

	- The CPS transform never adds calls that indirect through the `$call`
	  function.  Instead, all functions are inlined.
	- Neither the CPS nor the TRAMPOLINE transform add "guard" statements to the
	  fronts of functions
	- The TRAMPOLINE transform assumes that all functions are in trampolined style
	  and that therefore all function calls should be transformed to thunk returns;
	  the `$trampolineResult` function is never called.

Most of this removed functionality will be added again in the RUNTIME
transformation in stage 2.

I've added some "administrative" source-element types (CONTINUATION-FUNCTION and
THUNK-FUNCTION, which are both children of FUNCTION-EXPRESSION) to make it
easier for the RUNTIME transformation to treat these special functions specially
(eg, it doesn't make sense to assume that a continuation is in CPS).  These
administrative types are never created by the parser, only by source
transformations, so they make it a bit more complicated to write unit tests.  As
a solution, I've added the TEST-TRANSFORM function to TEST-UTILS.  It does the
transformation that is specified, and then follows that transformation with a
REMOVE-ADMINISTRATIVES transformation which convert administrative source
elements to their parseable equivalents.  I've fixed a bunch of unit tests using
this function already.

darcs-hash:20060114205513-53808-a37f10f50541331f31fb79c6a4b1110767f5d70a