init-runtime-transformation
This patch adds a working runtime transformation.  The functionality that was
removed from the CPS and TRAMPOLINE transformations by the
remove-runtime-dependencies patch has been added back to the RUNTIME
transformation.

The call-style guards that were previously checking to make sure that a CPS
function was being called with a function as its first argument have been
greatly improved.  We now check that the function is being called with a
continuation as an argument; if it isn't, then we start a trampoline running.

This means that the way to get a trampoline loop running is simply to call a
transformed function as if it had not been transformed, and everything else
should just automagically work.

Indirect calls via $call (for calling functions that may not be transformed)
is now added by transforming existing inlined calls, rather than by deciding
at CPS-conversion time whether a call needs to be inlined or not.

The $id continuation now returns a boxed result, so it should be safe to assume
that all continuations do so (by induction; all continuations wind up calling
$id at the end of the chain).  That means that we can do away with
$trampolineResult, since $id is now the place where the boxing of final results
occurs.

darcs-hash:20060115234928-53808-123a2b06431ab3a6fa614ba2e17c592b711f8ebe