fix-continuation-trampolining
Continuations (and in fact all pre-existing function expressions passed to tail-
calls) were not being recursed into properly.  This led to "first order"
continuation calls where the whole chain of continuation calls is on the stack,
which is exactly the sort of thing that trampolining is meant to avoid!

Fixed the transformation and added a regression test.

darcs-hash:20060117053637-53808-dcb2c918213759650a67dadd47e8a63036299935