fix-asdf-system
The addition of the custom js-file type was causing ASDF to recompile
the world every time it loaded the JWACS system.

The problem was that, although OPERATION-DONE-P always returns T for
LOAD-OP on a js-file, there was no corresponding method for COMPILE-OP.
Which meant that OPERATION-DONE-P always returned NIL for COMPILE-OP
on js-files.

So, I think what was going on was something like this:
We'd ask a CL file if it needed to be recompiled, and it would say
"I dunno, let me check my ancestors."  So then we would call
OPERATION-DONE-P with a COMPILE-OP operation on the js-file which happens
to be at the very beginning of our serial dependencies, which would say
it wasn't done, so obviously the "dependent" CL file needs to be rebuilt also!

ARRRGGH!

So, the fix is easy.  We now have a method on OPERATION-DONE-P COMPILE-OP JS-FILE
that always returns T.

darcs-hash:20060123192939-53808-9cc817c7fa09e84e621004bf7909a99f8ad418b5