labelled-statement-overhaul
Previously we represented labelled statements as a separate source element. Eg,
a labelled while would be represented as a LABEL node whose NAME slot held the
name of the label and whose STATEMENT slot held a WHILE structure.

However, this makes life a pain in the ass for actually trying to do things with
labelled statements, because it's not easy to know the label of the current
statement while you are processing that statement.

So now, every source-element has a LABEL slot, which is accessed using the
function SOURCE-ELEMENT-LABEL.  So if you're processing a WHILE (let's say :)
and you want to know whether it has a label, you can just call
(SOURCE-ELEMENT-LABEL ELM), and you should be set.

The addition of this world-wide LABEL slot means that the source-elements that
already /had/ a LABEL slot need to have it renamed.  So, the following renamings
are part of this patch:

BREAK-STATEMENT-LABEL    --> BREAK-STATEMENT-TARGET-LABEL
CONTINUE-STATEMENT-LABEL --> CONTINUE-STATEMENT-TARGET-LABEL
CASE-CLAUSE-LABEL        --> CASE-CLAUSE-VALUE

Be careful!  Case clauses and break and continue statements still have a LABEL
slot, but it doesn't mean what it used to.  The TARGET-LABEL slot is the one
that matters for these guys. ie:

    label:
    break targetLabel;

darcs-hash:20060126053114-53808-cf64130a935cbf67b1f27068108d2e52ba740d1f