fix-bare-fn-call-explicitization
We were running into some crazy problems where "bare" function calls (eg:
echo('hello!');) were being explicitized into variable declarations followed by
pointless variable references.  Eg:

	var JW0 = echo('hello!');
	JW0;

This was because our notion of when we were in a nested context was never
"cancelled".  For example, in the following code:

	var x = function() { ... body ... };

even though the function expression is in a nested context, the body is _not_.

We now shadow *NESTED-CONTEXT* for lists of statements and bodies of control
constructs like `for` or `if` statements.

darcs-hash:20060131042939-53808-6e098d9180152ab65dfce7064c51ec74d46786d3