make-ugly-print-earlier
ugly-print.lisp defines a macro (WITH-ADDED-ENVIRONMENT) that is used in
cps-transformation.lisp, but it was actually later than cps-transformation.lisp
in the build order, which caused warnings about undefined functions wherever
WITH-ADDED-ENVIRONMENT was used in cps-transformations.  This patch moves
ugly-print earlier in the build order to get rid of these warnings.

darcs-hash:20060204193759-53808-6fe5fcfccb47ac2d2fcb074b4ca84247651eebcb