type-analysis-function-calls
We are now able to analyze the types of function parameters and return values.
There are some basic tests for this new behaviour.

Some of the results are a little spooky.  It helps to recall that what we
determine is a superset of all possible types that a value could have.  For
example, in:

	var x = foo(10);
	foo('str');
	foo();

	function foo(a)
	{
		return a;
	}

the type list for x will actually be [Number, String, undefined] rather than the
[Number] that a more stringent type-checker would return.  The reasoning that it
uses looks like this:

	1. `foo` can return a Number, a String, or an undefined
	2. `x` is set to foo's return value
	3. `x` must be either a Number, a String, or an undefined.

I think that aside from adding extra `null`s and `undefined`s in various
variables' types, this behaviour will tend not to be as off-putting as in the
above example for real code.

darcs-hash:20060207054518-53808-27d2d15f7068f09c03c787aff67362822efb2fed