type-analysis/properties-working-but-putrid
This patch adds property support to the type analyzer.

This version is horrible!  The graph traversal algorithms are quadratic and the
code is not nicely factored in any way.  Writing it was a good way to figure out
the problem space, though.  The factored version (I plan to work on factoring it
this weekend) should be a lot nicer.

Not all expression types are supported yet.  I don't think that there are any
gotchas hiding in the as-yet-unsupported elements, though, so I'll probably mop
up a lot of those implementations after the refactoring.

This patch includes a function that converts a type graph into a DOT graph
description (which Graphviz can then turn into a pretty picture).  It's amazing
how much this can help with debugging this stuff.  The function basically works,
but is also in pretty desperate need of factoring.

darcs-hash:20060211204616-53808-3e80e19d3accca2e2b54f7f658bac1d159f41692