add-TrivialHttpRequest-example
We are now able to compile and run an example written in jwacs that uses the
XMLHttpRequest object without a manual-fixup stage.

The TrivialHttpRequest example demonstrates a very simple usage of a continuation-resuming
XMLHttpRequest wrapper.  To run it:

  1. Process TrivialHttpRequest.jw using JW::PROCESS.  It will produce TrivialHttpRequest.js.
  2. Load TrivialHttpRequest.html.
  3. The page should show three rows of 'data' (loaded from dataRows.txt using XMLHttpRequest).

darcs-hash:20060219045036-53808-30c10a70f9509d14d2899b6d0c14fe70dc199c72