refactor-type-analysis-4
This is the final type-analysis refactoring.

After reading a technical report entitled "Set-Based Analysis for Full Scheme
and Its Use in Soft-Typing" by Flanagan and Felleisen, I became convinced that
there may not actually be a particularly efficient soft-typing algorithm.  So
I've decided to abandon complicated and not-obviously-correct optimizations in
favour of a relatively slow but easy-to-verify algorithm.

Type analysis proceeds in three phases: POPULATE, CONNECT, and CONSTRUCT.

In the POPULATE phase, we walk the source tree.  Each identifier (and certain
expressions) are added to the type-graph as value-nodes.  Assignment edges are
added for each assignment; these edges indicate that the values of the rvalue
must be a subset of the values of the lvalue.  Eg, for the following code

	var x = 10;
	var y = 'str';
	y = x;

`x` and `y` would both get value nodes.  There is an edge from `x` to the
`Number` type-node (because `x`'s values include numbers); from `y` to the
`String` type-node; and from `y` to `x` (to account for the assignment).

In the CONNECT phase, we walk the type-graph adding additional edges based on
inference rules.  For example, in the following code:

	function foo(a, b)
	{
	  return a + b;
	}

	var x = foo('a', 'b');

After POPULATE:
	1. `foo`'s value-node points to the `foo` type-node
	2. `x`'s value node points to the $ret property of `foo`'s value-node
	3. `foo`'s value-node's first argument-node points to `String`
	4. `foo`'s value-node's second argument-node points to `String`
	5. the `foo` type-node's $ret property points to an anonymous value-node which
	   points to `a` and `b`'s value node.

CONNECT adds edges from `a` and `b` to the first and second argument nodes of
`foo` (to account for the function call), and from `foo`'s value-node's $ret
property to `foo`'s type-node's $ret property.

In the COLLAPSE phase, we walk the type-graph again with two goals:
	1. To gather all of the property, argument, and return nodes at the type-nodes
	2. To collapse the edges of the graph in such a way that every value-node points
	   to each type-node to which it had a path, and to /nothing else/.

In the above example, after the COLLAPSE phase:
	1. `foo`'s value-node points to the `foo` type-node
	2. `x`'s value-node points to the `String` type-node
	3. `a` and `b`'s value nodes both point to the `String` type-node
	4. The `foo` type-node's $ret property points to the `String` type-node.

The collapsed form makes COMPUTE-TYPES almost trivial to write; we just gather
and combine the types of a target expression's component parts.

The CONNECT and COLLAPSE algorithms are both pretty slow.

CONNECT adds a node back to the work queue every time it adds a new outgoing
connection; that can happen up to |v| times, and each traversal might cause
|v-1| other nodes to be visited.  So it seems like CONNECT is in the
neighbourhood of O(|v|^3) (yes, cubed).  This is actually the same bound that is
given for the algorithm in the technical report, so that may be as good as we
can do.

COLLAPSE isn't much better; Because of cycle handling, each node might be
visited once per "pass" (there is one pass per node), giving a bound of O(|v|^2).

COLLAPSE can probably be improved, although there's not much point if CONNECT
can't be.

darcs-hash:20060222044623-53808-0025f6b90e0a73826384e864a0b5dfff7571df30