rename-node-structures
This path renames the type-graph node structures in type-analysis. What used to
be called TYPE-NODEs are now called VALUE-NODEs, and what used to be called
VALUE-NODES are now called LOCATION-NODES.

It's a little confusing (esp. the re-use of the VALUE-NODE name), but I think
it's worthwhile, because it makes the names correspond a bit more closely to the
actual meanings.

In an upcoming enhancement (possibly tomorrow) I'm going to change the way we
find the value-node that a given constructor returns.  This should make
prototype handling a lot more straightforward, and it will esp. make life easier
with Object and Array values (which currently all map down to the same Object or
Array value-node, which is a little /too/ imprecise).

darcs-hash:20060224050801-53808-fdf77e92c86a69bec19d2039ba78be702692193f