clarify-prototype-handling
This checkin clarifies the relationship between function values and prototype
values.

Previously, a function Foo would have a location-node named Foo that pointed to
a value-node named Foo.  Values constructed using `new` that specified Foo as
the constructor would also be pointed to that same value-node.

However, this was incorrect, because it didn't distinguish between properties
that were set on Foo versus properties that were set on Foo.prototype.  ie, if
`Foo.prototype.bar = 10`, then every `new Foo` value will also have a bar
property with an initial value of 10.  However, if `Foo.baz = 10`, then that
should have no effect on `new Foo` values.  Under the old scheme, the two cases
were treated identically.

Under the new scheme:
  - location-node Foo points to value-node Foo
  - Foo has a "prototype" property-node which points to a Foo-proto value-node
  - instances of Foo point to Foo-proto instead of to Foo

This patch introduces the explicit concept of "exemplar values".  Basically, we
are treating each instance of Foo as being a reference to Foo.prototype; this
introduces some imprecision (since all instances are assumed to have the same
properties and values, which is not always true).

We make an exception for Array and Object values; we construct a new instance
value for each instance of these two types, because they are likely to vary
significantly in terms of which properties they have.

darcs-hash:20060225223948-53808-b87afa5601478433c8d0a87149bae08ed8027edf