refactor-prototype-handling-1
We were creating the value-node for prototype values in two different places
depending on whether we were in the POPULATE or CONNECT phase.  The only reason
this was necessary was due to a bug in CONNECT's handling of properties[1]. This
patch fixes the bug and cleans up our prototype handling by only ever creating
the prototype's value-node from the constructor-function's value node.

[1] The bug had two parts.  Firstly, we weren't properly queueing nodes for re-
processing when we added new outgoing edges, and secondly we weren't queueing
nodes that were created during CONNECT (specifically, the nodes for built-in
types such as `undefined`) for processing at all.

darcs-hash:20060225234435-53808-402b6688e649a0d70ccba46abf00625ed405a7de