constructor-args-handling
Type-analysis wasn't properly handling arguments passed to constructor
functions.  This patch changes the NEW-EXPR method of POPULATE-NODES to do the
proper assignment-edge and min-call-arity handling for constructor arguments.

This patch also updates handling for built-in types; we now call
SETUP-FUNCTION-NODE for built-in types (String, etc.) when their nodes are
requested using GET-LOCATION-NODE, instead of promiscuously setting up a function
node for each constructor argument to a NEW-EXPR (which resulted in extraneous
function nodes lying around).

One consequence is that a constructor must either be a function that exists or a
built-in type for new-expr to wind up correctly inferring types.  That seems
only fair to me; code that doesn't satisfy that requirement will result in a
runtime error anyway.

darcs-hash:20060226230206-53808-f14f3fb862397af5e81336f3161f519563163818