null-and-undefined-are-not-constructors
The clarifications continue!

`null` and `undefined` are builtin /values/, unlike `String` and `Number` and so
on, which are built-in /constructors/.  So we don't want to set up the
`undefined` "function", because it isn't a function.

So I've changed how we set up built-ins to correctly distinguish these two
cases.  COMPUTE-TYPES now gets its `undefined` type by getting the location-node
for `undefined` and returning its (single) assignment, rather than grabbing an
instance constructed by the `undefined` constructor (which is meaningless).  And
GET-LOCATION-NODE sets up a function for the built-in constructors (`String` et
al), but only a single non-function value-node for `null` and `undefined`.

darcs-hash:20060227003327-53808-f00ac9142a1e5324021604025d1eb98104c91e4a