more-efficient-cps-if-statements
The original implementation of the cps transformation for if-statements
duplicated the statement-tail in the then and else clauses, which in theory
could cause the size of the transformed code to be exponential in the number of
if statements in the original code.  This change causes us to place the
statement-tail into a named continuation (as in the `continue` and `break`
statement transformations), which is resumed by both the then and if clauses.

darcs-hash:20060319215837-53808-31a2981296d3c7fe9ec9782e8b5bbfdd32b69d5c