cps-switch-support
The cps transformation now supports switch statements.  We use the technique of
generating a named continuation which each branch resumes, similarly to if
statements and while loops.

We deal with "fall-through" by absorbing the statements of the following clause
for each "unterminated" clause:

	case 10:
	  doSomething();
	case 20:
	case 30:
	  doSomethingElse();
	  break;
	default:

	=== becomes ===>

	case 10:
	  doSomething();
	  doSomethingElse();
	  break;
	case 20:
	case 30:
	  doSomethingElse();
	  break;
	default:
	  break;

Note that we treat "null clauses" (ie, those with no statements in their own
body) as a special case by continuing to allow them to fall through to the next
clause.  If the final clause is unterminated, we slap a break onto the end (to
ensure that the tail continuation gets resumed).

To support this change, I've extended EXPLICIT-RETURN-P to allow the caller to
specify which types of statement terminate a group of statements (since the
switch code cares about `break` and `continue` in addition to `return`, `throw`,
`suspend`, and `resume`. EXPLICIT-RETURN-P is now a wrapper around the more
general EXPLICITLY-TERMINATED-P.

darcs-hash:20060321044713-53808-85898609b462dbbaf9fa682dc063386004a64980