nested-switch-support
Turns out switch statements can be labelled (so that you can `break` out of
multiple switches at once, just like in loops).  So I've extended the switch
transformation to bind the switch's break continuation in the environment for
labelled switches (so that `break` statements with a target call the right
continuation).

Also, the nesting tests exposed a bug with the break continuation declaration.
We can't use a function-decl to create the break continuation, because that is
not a statement, and the switch statement may occur in contexts where only
statements (as distinct from source-elements) are allowed (eg, inside another
switch statement).  So now we set a variable to a function-expression instead.

Also, we need to make sure to transform the body of the break continuation
outside the scope of the *nearest-break* binding; otherwise nested null breaks
won't go to the right continuation.

It looks like the cps transformation for while statements has the latter two
nesting-related errors; I'll fix those (and add some tests) in a separate patch.

darcs-hash:20060322053000-53808-a022ff2074afb22c5fdacc00b25887730fc5c867