more-efficient-cps-if-statements-2
This patch adds further optimizations for cps transforming if statements.

The first pass at optimizing the cps transformation for if statements (in patch
more-efficient-cps-if-statements) caused us to /always/ generate a named
continuation for every if statement no matter what.  That's not ideal, because
in lots of common cases that actually makes the code longer (not to mention more
confusing, and more closure-happy and therefore probably slower).

We only really need a named continuation if both branches of the if statement
are non-null and unterminated.  This patch does a more sophisticated termination
check to determine whether to generate a named continuation or not.

darcs-hash:20060325203658-53808-b7ce210156dc668bfbb425986479230f88e0d908