fix-object-literal-string-field-names
When we converted to storing field names as STRING-LITERALs instead of
IDENTIFIERs, we didn't update the pretty-printer, so we were printing object
literals in this form:

	{"field": 10, "field2", 20}

That's syntactically correct but kind of ugly.  And this is a pretty-printer.
So I've added a special case to print string field names in the identifier style:

	{field: 10, field2: 20}

darcs-hash:20060402053432-53808-43ccf2a6dccdaf1b995f97747451ce98485cd617