named-continuation->labelled-continuation
Just a nomenclature change.  I've referred in a few places in the code to "named
continuations" (where we basically create a continuation and assign it to a
variable to be referred to by multiple locations in the code).  I think that
it's a little clearer to refer to these as "labelled continuations" instead,
since they are a little like the "labels" that you produce when compiling a
functional language.

On the other hand, maybe it's actually an abuse of terminology; we won't really
know until we try it for a while. :)

darcs-hash:20060404021628-53808-360931dc0072d74ae7f14e37ecc12904476fc30a