refactor-loop-transformation
The loop transformation was failing in the presence of for loops whose body only
contained a single statement (since it assumed that the body of the loop would
be a STATEMENT-BLOCK, which is true only when there are multiple statements in
the body).

Fixed by refactoring the loop transformation to use SINGLE-STATEMENT to
construct its output element (instead of calling MAKE-STATEMENT-BLOCK directly
on an appended-together list of lists).  This makes the code more readable and
also more robust.

darcs-hash:20060404025635-53808-cda9723b01d6c18d5d0821f47a3a50d9fa4f29dc