introduce-defelement
This patch introduces the DEFELEMENT macro, which replaces the previous
shadowing DEFSTRUCT macro.  There was some sort of funny compilation edge-case
where, if things got compiled in the wrong order, we'd run into package-lock
issues with the DEFSTRUCT macro.  That was probably solvable by moving the
DEFSTRUCT macro into its own file or something, but frankly it's not worth the
bother.

Instead, we now have the exact same macro, except with a name that does /not/
conflict with a CL-package symbol, which hopefully will resolve the issue.

(the point of the DEFELEMENT macro, as with the DEFSTRUCT macro that preceded
it, is to add MAKE-LOAD-FORM calls for each new source-element struct so that
we can use literal structure values in variables that will be dumped to fasls
under SBCL, which doesn't provide default load-forms for structs or classes).

darcs-hash:20060405035247-53808-d53bc310bdda5bbec9699ee657c2f2b91c1fbe39