add-import-syntax
Adds syntax for import declarations to jwacs.

An import declaration is a source-element, not a statement.  The intention is
that they should only appear at the beginning of the file, although I don't
enforce that in any way.

The syntax is:

    import [<type_specifier>] "<uri>";

Eg:
    import js "/prototype.js";
    import jw "lib1.jw";
    import "lib2.jw";

File type is intended to be inferred from the extension in the absence of an
explicit type specifier.

I'm starting to think that the only sort of imports that make sense are actually
js or jw files.  CSS etc. support should probably be in the template file rather
than the script files.

darcs-hash:20060423180141-53808-34f2267eca1112125ad86ce7b558e095726a3dac