add-build-app
We now provide the BUILD-APP function, which will generate an entire jwacs
application from a single source file (potentially).  A template file can be
used if/when the default wrapper html file becomes inadequate.

This version includes support for importing Javascript or jwacs sources by URI,
but not for importing HTML or CSS.  I think that the template approach is a
better way to customize the actual application HTML file than a limited import
scheme; inevitable somebody will be unhappy with the limitations of a import-
only based scheme.  (In fact, the motivation for the switch was that /I/ was
unhappy with it; lang-tests is much nicer as a single HTML page with a bit of
scripting than as a frameset with bootstrapper etc.)

darcs-hash:20060423215853-53808-c80d6b4bef3c63f1edebf1a5ed197f3cb065823c