explicitize-new-exprs
We now properly explicitize new expressions.  It's important to treat the
constructor arguments properly  (so that we aren't trying to embed
return statements into expressions further down the road), so this:

  new Foo(bar() + 1)

now becomes

  var JWn = bar();
  new Foo(JWn);

It's also important to treat new expressions the same as functions calls
generally (ie, we should be proxying new expressions that occur within
expressions):

  var foo = bar(new Baz(100));

now becomes

  var JWn = new Baz(100);
  var foo = bar(JWn);

The reason for this is that a new expression is actually an implicit function
call to the constructor.  In a subsequent patch I plan to modify the runtime
transformation to transform these new expressions into explicit function calls
to a runtime function (`$new`, probably) that will allow us to correctly support
constructors that use continuation features.

darcs-hash:20060606152953-53808-2acff1bcfcde8bc91555f85c417391e5bf69e65b