fix-uniquification-case-sensitivity
We were using EQUALP to find identifier bindings during uniquification, which is
a problem since identifier names are case-sensitive but EQUALP is case-insensitive.

This patch switches from EQUALP to EQUAL (which is case-sensitive) and adds a
couple of tests for case-sensitivity.  It also fixes a couple of tests that were
incorrectly expecting case-insensitive behaviour.

darcs-hash:20060606163958-53808-95a5d6ad546589143bf76c1421ba412b873cedd9