trampoline/runtime:new-expr-support
The trampoline and runtime transformations now support our sneaky scheme for
converting new expressions into explicit constructor calls.  Supporting
functions in jw-runtime.js have also been added (`$makeBlank`, `$new0`, and
`$new`).

The sneaky scheme works like this:

1) CPS transform treats new expressions like regular function calls.
   So `new Hello(10);` becomes `return new Hello(function(JWn) { ... }, 10);`.

2) Trampoline transform treats new expressions like regular function calls
   Basically this means that we wrap them in a thunk instead of returning them
   as a boxed value.

3) Runtime transformation converts new expressions into calls to the runtime
   `$new0` function.  (usually; for many-arg constructors it'll turn into a
   call to `$new` instead)

The magic is all in the `$new0` function.  The basic idea is to call the
constructor function ourselves instead of letting the interpreter do it for us.
There's a complication, though, in that we still need to construct an object
/with the correct implicit prototype/ before we can call the constructor on it.

We do this by creating a "blank constructor".  A blank constructor is just a
function that does nothing.  However, we set its prototype field to equal the
prototype field of the "real" constructor!  Eg:

    var blank = function() {};
    blank.prototype = ctor.prototype;
    var obj = new blank;

After the above runs, `obj` has the same (implicit and not directly accessible)
prototype as it would have had if we had created it using `new ctor`.  The
difference is that none of the initialization code contained in `ctor` has been
run; we have a pristine, uninitialized object with the correct prototype.

Once we have that, it's pretty straightforward for `$new0` to call the real
constructor as a method call on `obj` and return the result.

==============
FOR DISCUSSION
==============

I'm a little uncomfortable about the level of coupling that this change
introduces.  Up until now, code that came out of a transformation was always
valid if the code that went in had been.  But now, new expressions are invalid
after the cps transformation, and they stay that way until the runtime
transformation finally fixes them up.

darcs-hash:20060607172002-53808-9f4a0a582fe03b6918fe66a636fef9f526fd0407